# 8352536: Add overloads to parse and build class files from/to MemorySegment