… src/java.base/share/classes/jdk/internal/classfile/impl/ClassFileImpl.java
… src/java.base/share/classes/jdk/internal/classfile/impl/DirectClassBuilder.java
… test/jdk/jdk/classfile/BuildAndParseBuffersAndSegments.java
… test/micro/org/openjdk/bench/jdk/classfile/MemorySegmentBenchmark.java