Provide method overloads to the ClassFile interface of the java.lang.classfile API which allow parsing of classes found in memory segments, as well as allowing built class files to be output to them.

<!-- Anything below this marker will be automatically updated, please do not edit manually! -->
---------
### Progress
- [ ] Change must be properly reviewed (1 review required, with at least 1 [Reviewer](https://openjdk.org/bylaws#reviewer))
- [x] Change must not contain extraneous whitespace
- [x] Commit message must refer to an issue
- [ ] Change requires CSR request [JDK-8352562](https://bugs.openjdk.org/browse/JDK-8352562) to be approved

### Issues
 * [JDK-8352536](https://bugs.openjdk.org/browse/JDK-8352536): Add overloads to parse and build class files from/to MemorySegment (**Enhancement** - P4)
 * [JDK-8352562](https://bugs.openjdk.org/browse/JDK-8352562): Add overloads to parse and build class files from/to MemorySegment (**CSR**)


### Reviewing
<details><summary>Using <code>git</code></summary>

Checkout this PR locally: \
`$ git fetch https://git.openjdk.org/jdk.git pull/24139/head:pull/24139` \
`$ git checkout pull/24139`

Update a local copy of the PR: \
`$ git checkout pull/24139` \
`$ git pull https://git.openjdk.org/jdk.git pull/24139/head`

</details>
<details><summary>Using Skara CLI tools</summary>

Checkout this PR locally: \
`$ git pr checkout 24139`

View PR using the GUI difftool: \
`$ git pr show -t 24139`

</details>
<details><summary>Using diff file</summary>

Download this PR as a diff file: \
<a href="https://git.openjdk.org/jdk/pull/24139.diff">https://git.openjdk.org/jdk/pull/24139.diff</a>

</details>
<details><summary>Using Webrev</summary>

[Link to Webrev Comment](https://git.openjdk.org/jdk/pull/24139#issuecomment-2741520709)
</details>