@asotona 2025-03-27
Class-File API passed through several significant weight reductions in its history (on the API surface as well as under the skin). Some of the steps were critical to avoid JDK bootstrap performance regression.

Before deciding about this non-trivial addition to the Class-File API I would like to see some measurable benefits  and performance impact on the JDK bootstrap.