@asotona 2025-03-27
I'm sorry, but I'm still missing the benefits parh of this change.

The costs are high. The API grows, implementation splits into multiple parallel branches in ClassFileImpl and DirectClassBuilder. 
A new implementation class BuildAndParseBuffersAndSegments appears.
Parallel implementations add a new dimension to the test coverage matrix.
There is also still a lot of code related to ByteBuffer.

I'm not convinced the API should be extended with the MemorySegment overrides.