@asotona 2025-03-27
As the parsing benefits are none due to the memory copy.
May I see at least the benefits of direct writing to memory mapped files on some benchmark?