@dmlloyd 2025-03-27
> As the parsing benefits are none due to the memory copy.

The main benefit on the parsing is twofold: reducing the user's boilerplate, and the future possibility of further optimizing this case. At present the performance is the same as making the user do the copy.

> May I see at least the benefits of direct writing to memory mapped files on some benchmark?

Sure, I can put together a benchmark which compares the throughput with and without the copy and maybe also compares the allocation rate. I hope that would be sufficient to capture whatever there is to capture.